+++ src/Data/Managers/GetAgents/QueryManager.ts
@@ -4,13 +4,10 @@ import {
   stringifyObjectOrUndefined,
   StateHelperWithEnv,
 } from "@/Core";
-import {
-  getPaginationHandlers,
-  PrimaryContinuousQueryManagerWithEnv,
-} from "@/Data/Managers/Helpers";
+import { getPaginationHandlers, QueryManager } from "@/Data/Managers/Helpers";
 import { getUrl } from "./getUrl";
 
-export class GetAgentsQueryManager extends PrimaryContinuousQueryManagerWithEnv<"GetAgents"> {
+export class GetAgentsQueryManager extends QueryManager.ContinuousWithEnv<"GetAgents"> {
   constructor(
     apiHelper: ApiHelper,
     stateHelper: StateHelperWithEnv<"GetAgents">,
