+++ src/Data/Managers/GetDesiredStateDiff/QueryManager.ts
@@ -1,8 +1,8 @@
 import { identity } from "lodash-es";
 import { ApiHelper, StateHelperWithEnv } from "@/Core";
-import { PrimaryOneTimeQueryManagerWithEnv } from "@/Data/Managers/Helpers";
+import { QueryManager } from "@/Data/Managers/Helpers";
 
-export class GetDesiredStateDiffQueryManager extends PrimaryOneTimeQueryManagerWithEnv<"GetDesiredStateDiff"> {
+export class GetDesiredStateDiffQueryManager extends QueryManager.OneTimeWithEnv<"GetDesiredStateDiff"> {
   constructor(
     apiHelper: ApiHelper,
     stateHelper: StateHelperWithEnv<"GetDesiredStateDiff">
