+++ src/Data/Managers/GetParameters/QueryManager.ts
@@ -4,13 +4,10 @@ import {
   StateHelperWithEnv,
   stringifyObjectOrUndefined,
 } from "@/Core";
-import {
-  getPaginationHandlers,
-  PrimaryContinuousQueryManagerWithEnv,
-} from "@/Data/Managers/Helpers";
+import { getPaginationHandlers, QueryManager } from "@/Data/Managers/Helpers";
 import { getUrl } from "./getUrl";
 
-export class GetParametersQueryManager extends PrimaryContinuousQueryManagerWithEnv<"GetParameters"> {
+export class GetParametersQueryManager extends QueryManager.ContinuousWithEnv<"GetParameters"> {
   constructor(
     apiHelper: ApiHelper,
     stateHelper: StateHelperWithEnv<"GetParameters">,
