+++ src/Data/Managers/Helpers/QueryManager/types.ts
