+++ src/Data/Managers/Resources/QueryManager.ts
@@ -1,11 +1,8 @@
 import { Scheduler, Resource, ApiHelper, StateHelperWithEnv } from "@/Core";
-import {
-  getPaginationHandlers,
-  PrimaryContinuousQueryManagerWithEnv,
-} from "@/Data/Managers/Helpers";
+import { getPaginationHandlers, QueryManager } from "@/Data/Managers/Helpers";
 import { getUrl } from "./getUrl";
 
-export class ResourcesQueryManager extends PrimaryContinuousQueryManagerWithEnv<"GetResources"> {
+export class ResourcesQueryManager extends QueryManager.ContinuousWithEnv<"GetResources"> {
   constructor(
     apiHelper: ApiHelper,
     stateHelper: StateHelperWithEnv<"GetResources">,
